requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
feedparser==6.0.8
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
